requirements.txt
-e .
aws-cdk.aws-lambda==1.84.0
aws-cdk.aws_events==1.84.0
aws-cdk.aws_events_targets==1.84.0
aws-cdk.core==1.84.0
